+++ assets/mixins/form.mixin.js
@@ -14,11 +14,12 @@ formMixin.resetForm = function formMixinResetForm(formName) {
   var inputs = this.getInputs(formName);
 
   _.forEach(inputs, function formMixinResetFormRefsLoop(input, key) {
+    /* istanbul ignore else */
     if (this.refs[key] && this.refs[key].validator) {
       this.refs[key].validator.reset();
 
-      if(this.refs[key].props.validateOnLoad === true) {
-        this.refs[key].validator.validate(this.refs[key].cleanValue(this.refs[key].props[input.valueProperty]));
+      if (this.refs[key].props.validateOnLoad === true) {
+        this.refs[key].validator.validate(this.refs[key].cleanValue(resetData[formName][key]));
       }
 
       this.refs[key].forceUpdate();
@@ -30,6 +31,7 @@ formMixin.validateForm = function formMixinValidateForm(formName) {
   var inputs = this.getInputs(formName);
 
   _.forEach(inputs, function formMixinResetFormInputLoop(input, key) {
+    /* istanbul ignore else */
     if (this.refs[key] && this.refs[key].validator) {
       this.refs[key].validator.validate(this.refs[key].cleanValue(this.refs[key].props[input.valueProperty]));
       this.refs[key].forceUpdate();